Moved Loc to common resources
