+++ src/AppInstallerCLICore/Resources.h
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 #pragma once
-#include <winget/LocIndependent.h>
 #include <winget/Resources.h>
 
 #include <winrt/Windows.ApplicationModel.Resources.h>
@@ -11,6 +10,9 @@
 namespace AppInstaller::CLI::Resource
 {
     using AppInstaller::StringResource::StringId;
+    using AppInstaller::Resource::LocString;
+    using AppInstaller::Resource::Loader;
+    using AppInstaller::Resource::ResourceOpenException;
 
     // Resource string identifiers.
     // This list can mostly be generated by the following PowerShell:
@@ -403,36 +405,6 @@ namespace AppInstaller::CLI::Resource
         WINGET_DEFINE_RESOURCE_STRINGID(WordArgumentDescription);
     };
 
-    // A localized string
-    struct LocString : public Utility::LocIndString
-    {
-        LocString() = default;
-
-        LocString(StringId id);
-
-        LocString(const LocString&) = default;
-        LocString& operator=(const LocString&) = default;
-
-        LocString(LocString&&) = default;
-        LocString& operator=(LocString&&) = default;
-    };
-
-    // Utility class to load resources
-    class Loader
-    {
-    public:
-        // Gets the singleton instance of the resource loader.
-        static const Loader& Instance();
-
-        // Gets the the string resource value.
-        std::string ResolveString(std::wstring_view resKey) const;
-
-    private:
-        winrt::Windows::ApplicationModel::Resources::ResourceLoader m_wingetLoader;
-
-        Loader();
-    };
-
     // Fixed strings are not localized, but we use a similar system to prevent duplication
     enum class FixedString
     {
@@ -440,16 +412,6 @@ namespace AppInstaller::CLI::Resource
     };
 
     Utility::LocIndView GetFixedString(FixedString fs);
-
-    struct ResourceOpenException : std::exception
-    {
-        ResourceOpenException(const winrt::hresult_error& hre);
-
-        const char* what() const noexcept override { return m_message.c_str(); }
-
-    private:
-        std::string m_message;
-    };
 }
 
 inline std::ostream& operator<<(std::ostream& out, AppInstaller::CLI::Resource::StringId si)
