+++ src/AppInstallerCommonCore/Public/winget/Resources.h
@@ -5,6 +5,7 @@
 
 #include <string>
 #include <vector>
+#include "LocIndependent.h"
 
 using namespace std::string_view_literals;
 
@@ -55,5 +56,45 @@ namespace AppInstaller
         // Resource data is valid as long as the binary is loaded.
         std::pair<const BYTE*, size_t> GetResourceAsBytes(int resourceName, int resourceType);
         std::pair<const BYTE*, size_t> GetResourceAsBytes(PCWSTR resourceName, PCWSTR resourceType);
+
+        // A localized string
+        struct LocString : public Utility::LocIndString
+        {
+            LocString() = default;
+
+            LocString(StringResource::StringId id);
+
+            LocString(const LocString&) = default;
+            LocString& operator=(const LocString&) = default;
+
+            LocString(LocString&&) = default;
+            LocString& operator=(LocString&&) = default;
+        };
+
+        // Utility class to load resources
+        class Loader
+        {
+        public:
+            // Gets the singleton instance of the resource loader.
+            static const Loader& Instance();
+
+            // Gets the the string resource value.
+            std::string ResolveString(std::wstring_view resKey) const;
+
+        private:
+            winrt::Windows::ApplicationModel::Resources::ResourceLoader m_wingetLoader;
+
+            Loader();
+        };
+
+        struct ResourceOpenException : std::exception
+        {
+            ResourceOpenException(const winrt::hresult_error& hre);
+
+            const char* what() const noexcept override { return m_message.c_str(); }
+
+        private:
+            std::string m_message;
+        };
     }
 }
