+++ src/AppInstallerCommonCore/Resources.cpp
@@ -4,6 +4,7 @@
 #include "winget/Resources.h"
 #include "Public/AppInstallerLogging.h"
 #include "Public/AppInstallerStrings.h"
+#include "Public//AppInstallerErrors.h"
 
 namespace AppInstaller::Resource
 {
@@ -54,5 +55,45 @@ namespace AppInstaller::Resource
     {
         auto resourceData = GetResourceData(resourceName, resourceType);
         return std::make_pair(reinterpret_cast<BYTE*>(resourceData.first), resourceData.second);
+    }
+
+    LocString::LocString(StringResource::StringId id) : Utility::LocIndString(Loader::Instance().ResolveString(id)) {}
+
+    const Loader& Loader::Instance()
+    {
+        static Loader instance;
+        return instance;
+    }
+
+    Loader::Loader() : m_wingetLoader(nullptr)
+    {
+        try
+        {
+            // The default constructor of ResourceLoader throws a winrt::hresult_error exception
+            // when resource.pri is not found. ResourceLoader::GetForViewIndependentUse also throws
+            // a winrt::hresult_error but for reasons unknown it only gets catch when running on the
+            // debugger. Running without a debugger will result in a crash that not even adding a
+            // catch all fix. To provide a good error message we call the default constructor
+            // before calling GetForViewIndependentUse.
+            m_wingetLoader = winrt::Windows::ApplicationModel::Resources::ResourceLoader();
+            m_wingetLoader = winrt::Windows::ApplicationModel::Resources::ResourceLoader::GetForViewIndependentUse(L"winget");
+        }
+        catch (const winrt::hresult_error& hre)
+        {
+            // This message cannot be localized.
+            AICLI_LOG(CLI, Error, << "Failure loading resource file with error: " << hre.code());
+            throw ResourceOpenException(hre);
+        }
+    }
+
+    std::string Loader::ResolveString(
+        std::wstring_view resKey) const
+    {
+        return Utility::ConvertToUTF8(m_wingetLoader.GetString(resKey));
+    }
+
+    ResourceOpenException::ResourceOpenException(const winrt::hresult_error& hre)
+    {
+        m_message = "Could not open the resource file: " + GetUserPresentableMessage(hre);
     }
 }
